improve spacing, error messages
Error messages should include relevant info about what failed, and use
warning log levels for misuse and error levels when there are critical
failures, such as lack of support.

ideally, the build system should also check that the kernels supports
resume, that way it should never fail at runtime.
This shouldn't be that serious, because if the system doesn't support
resume, it won't hibernate, so there is no risk of data loss if it fails
to resume later.

Spacing in the if - echo section is a bit confusing but visually lines
up. May be altered by dedent later.

Signed-off-by: Zen <z@pyl.onl>